+++ bootloader/dnc-commonlib.h
@@ -73,7 +73,7 @@ extern char *microcode_path;
 extern int disable_smm;
 extern int renumber_bsp;
 extern int mem_offline;
-extern int trace_buf_size;
+extern u64 trace_buf_size;
 extern int verbose;
 
 extern u16 shadow_rtbll[7][256];
